% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roussel_et_al_2017.R
\name{hmax_error_local}
\alias{hmax_error_local}
\title{Max height error estimation local}
\usage{
hmax_error_local(las, res)
}
\arguments{
\item{las}{LAS object from lidR}

\item{res}{scalar. The resolution at which the estimation must be performed}
}
\value{
A list with 2 data.frame. The first contains the normalized histogram of probabilities
(figure 13) and the second contains the estimated errors (figure 14). See references.
}
\description{
This function evaluates, for a given pixel size the expected error made on the metrics hmax as a
function of the pulse density according to Roussel et al (2007) (see references). The metric hmax
is always underestimated because discrete laser pulses are statistically unlikely to hit the very
top of the canopy. The highest laser hit is necessarily below the theoretically highest point. How
much? It depends on the shape/complexity of the canopy and the sampling density. By evaluating the
shape of the canopy and using probability theory, it is possible to estimate the expected error
made as a function of the pulse density (see references).
}
\examples{
library(lidR)
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
las = readLAS(LASfile)

shapefile_dir <- system.file("extdata", "lake_polygons_UTM17.shp", package = "lidR")
lakes = sf::st_read(shapefile_dir)

las = classify_poi(las, LASWATER, roi = lakes)
las = filter_poi(las, Classification != LASWATER)

res = 20
ans = hmax_error_local(las, res)

library(ggplot2)
# plot the histogram of probabilities (fig 13)
ggplot(ans$shape) +
 aes(x = height, y = probability) +
 geom_line() +
 coord_flip() +
 ylab("probability") +
 ylim(0,0.02) +
 xlab("Height") +
 theme_bw()

# plot dependence of bias on pulse density (fig 14)
ggplot(ans$error) +
 aes(x = density, y = error) +
 geom_line() +
 scale_x_log10(breaks = c(1,2,3,4,5,10,20,30,50)) +
 scale_y_continuous(limits=c(-1.3,0), breaks =seq(-1.25,0,0.25)) +
 xlab("Pulse density") +
 ylab("Theorical error") +
 ggtitle(paste("Abacus computed for", res*res, "square meters")) +
 theme_bw()

}
\references{
Roussel, J.-R., Caspersen, J., Béland, M., Thomas, S., & Achim, A. (2017). Removing
bias from LiDAR-based estimates of canopy height: Accounting for the effects of pulse density and
footprint size. Remote Sensing of Environment, 198, 1–16. https://doi.org/10.1016/j.rse.2017.05.032
}
