% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roussel_et_al_2017.R
\name{hmax_error_metric}
\alias{hmax_error_metric}
\title{Error metric}
\usage{
hmax_error_metric(X, Y, Z, res, at)
}
\arguments{
\item{X, Y, Z, res}{Internal use.}

\item{at}{numeric. Densities to estimate}
}
\description{
Exported for internal use only
}
